create lib/hdtype.lua, move hd_type code into hdtype.lua
